Move categoryPermission to DiscussionModel
