+++ applications/vanilla/models/class.vanillamodel.php
@@ -8,8 +8,6 @@
  * @since 2.0
  */
 
-use Vanilla\Exception\PermissionException;
-
 /**
  * Introduces common methods that child classes can use.
  *
@@ -56,26 +54,5 @@ abstract class VanillaModel extends Gdn_Model {
         $isUserSpamming = $this->isUserSpamming($session->User->UserID, $storageObject);
 
         return $isUserSpamming;
-    }
-
-    /**
-     * Verify the current user has a permission in a category.
-     *
-     * @param string|array $permission The permission slug(s) to check (e.g. Vanilla.Discussions.View).
-     * @param int $categoryID The category's numeric ID.
-     * @throws PermissionException if the current user does not have the permission in the category.
-     */
-    public function categoryPermission($permission, $categoryID) {
-        $category = CategoryModel::categories($categoryID);
-        if ($category) {
-            $id = $category['PermissionCategoryID'];
-        } else {
-            $id = -1;
-        }
-        $permissions = (array)$permission;
-
-        if (!Gdn::session()->getPermissions()->hasAny($permissions, $id)) {
-            throw new PermissionException($permissions);
-        }
     }
 }
